出品者による削除（非同期）

+++ app/assets/javascripts/comment.js
@@ -1,30 +1,39 @@
 $(function(){
   function new_comment(comment_data){
-    var html_comment = 
+    if (comment_data.item_user.id == comment_data.user_id){
+      var html = 
       `
       <div class="comment-Me">
         <div class="comment-content">
           ${comment_data.comment}
-          `
-    // if (comment_data.user_of_item.id == comment_data.id_of_user_of_comment){
-    var html_delete =
-          `
+          <div class="comment_create_at">
+            ${comment_data.created_at}
+          </div>
           <div class="comment-delete">
             <a rel="nofollow" data-method="delete" href="/comments/${comment_data.id}">削除する</a>
           </div>
         </div>
-        `
-        // }else{
-        //   var html_delete = ``
-        // };
-
-    var html_name =`
         <div class="comment-user-name">
           ${comment_data.user_nickname}
         </div>
-      </div>`
+      </div> 
+      `
+    }else{
+      var html = 
+      `<div class="comment-Me">
+        <div class="comment-content">
+        ${comment_data.comment}
+        <div class="comment_create_at">
+          ${comment_data.created_at}
+        </div>
+        </div>
+        <div class="comment-user-name">
+        ${comment_data.user_nickname}
+        </div>
+      </div>`   
+          };
 
-    var html = html_comment + html_delete + html_name;
+
     return html;
 
   }
